Split CLI integration tests by feature
